@@ -18,7 +18,7 @@
  */
 
 #include <folly/Conv.h>
-#include <folly/Logging.h>
+#include <folly/GLog.h>
 #include <folly/io/Cursor.h>
 #include <google/protobuf/io/coded_stream.h>
 #include <google/protobuf/io/zero_copy_stream_impl_lite.h>
