@@ -17,13 +17,13 @@
  *
  */
 
-#include <folly/Logging.h>
+#include <folly/logging/Logger.h>
 #include <folly/Memory.h>
 #include <folly/futures/Future.h>
 #include <folly/io/async/EventBase.h>
 #include <folly/io/async/ScopedEventBaseThread.h>
 #include <gmock/gmock.h>
-#include <wangle/concurrent/IOThreadPoolExecutor.h>
+#include <folly/executors/IOThreadPoolExecutor.h>
 
 #include <chrono>
 #include <functional>
@@ -183,9 +183,9 @@ class MockAsyncConnection : public AsyncConnection,
  public:
   MockAsyncConnection(std::shared_ptr<ConnectionConfiguration> conn_conf,
                       std::shared_ptr<folly::HHWheelTimer> retry_timer,
-                      std::shared_ptr<wangle::CPUThreadPoolExecutor> cpu_executor,
-                      std::shared_ptr<wangle::IOThreadPoolExecutor> io_executor,
-                      std::shared_ptr<wangle::IOThreadPoolExecutor> retry_executor,
+                      std::shared_ptr<folly::CPUThreadPoolExecutor> cpu_executor,
+                      std::shared_ptr<folly::IOThreadPoolExecutor> io_executor,
+                      std::shared_ptr<folly::IOThreadPoolExecutor> retry_executor,
                       std::shared_ptr<RpcClient> rpc_client,
                       std::shared_ptr<AsyncRegionLocator> region_locator)
       : conn_conf_(conn_conf),
@@ -208,9 +208,9 @@ class MockAsyncConnection : public AsyncConnection,
   }
   std::shared_ptr<RpcClient> rpc_client() override { return rpc_client_; }
   std::shared_ptr<AsyncRegionLocator> region_locator() override { return region_locator_; }
-  std::shared_ptr<wangle::CPUThreadPoolExecutor> cpu_executor() override { return cpu_executor_; }
-  std::shared_ptr<wangle::IOThreadPoolExecutor> io_executor() override { return io_executor_; }
-  std::shared_ptr<wangle::IOThreadPoolExecutor> retry_executor() override {
+  std::shared_ptr<folly::CPUThreadPoolExecutor> cpu_executor() override { return cpu_executor_; }
+  std::shared_ptr<folly::IOThreadPoolExecutor> io_executor() override { return io_executor_; }
+  std::shared_ptr<folly::IOThreadPoolExecutor> retry_executor() override {
     return retry_executor_;
   }
 
@@ -225,9 +225,9 @@ class MockAsyncConnection : public AsyncConnection,
   std::shared_ptr<AsyncRpcRetryingCallerFactory> caller_factory_;
   std::shared_ptr<RpcClient> rpc_client_;
   std::shared_ptr<AsyncRegionLocator> region_locator_;
-  std::shared_ptr<wangle::CPUThreadPoolExecutor> cpu_executor_;
-  std::shared_ptr<wangle::IOThreadPoolExecutor> io_executor_;
-  std::shared_ptr<wangle::IOThreadPoolExecutor> retry_executor_;
+  std::shared_ptr<folly::CPUThreadPoolExecutor> cpu_executor_;
+  std::shared_ptr<folly::IOThreadPoolExecutor> io_executor_;
+  std::shared_ptr<folly::IOThreadPoolExecutor> retry_executor_;
 };
 
 template <typename CONN>
@@ -269,12 +269,12 @@ class MockRawAsyncTableImpl {
     auto f = promise_->getFuture();
     VLOG(1) << "calling rpc_call";
     rpc_call(rpc_client, loc, controller, std::move(req_converter(req, loc->region_name())))
-        .then([&, this, resp_converter](std::unique_ptr<PRESP> presp) {
+        .thenValue([&, this, resp_converter](std::unique_ptr<PRESP> presp) {
           VLOG(1) << "MockRawAsyncTableImpl#call succeded: ";
           RESP result = resp_converter(*presp);
           promise_->setValue(result);
         })
-        .onError([this](const exception_wrapper &e) {
+        .thenError([this](const exception_wrapper &e) {
           VLOG(1) << "entering MockRawAsyncTableImpl#call, exception: " << e.what();
           VLOG(1) << "entering MockRawAsyncTableImpl#call, error typeinfo: " << typeid(e).name();
           promise_->setException(e);
@@ -310,10 +310,10 @@ void runTest(std::shared_ptr<AsyncRegionLocatorBase> region_locator, std::string
   /* init region location and rpc channel */
   auto region_location = table->GetRegionLocation(row);
 
-  // auto io_executor_ = std::make_shared<wangle::IOThreadPoolExecutor>(4);
-  auto cpu_executor_ = std::make_shared<wangle::CPUThreadPoolExecutor>(4);
+  // auto io_executor_ = std::make_shared<folly::IOThreadPoolExecutor>(4);
+  auto cpu_executor_ = std::make_shared<folly::CPUThreadPoolExecutor>(4);
   auto io_executor_ = client.async_connection()->io_executor();
-  auto retry_executor_ = std::make_shared<wangle::IOThreadPoolExecutor>(1);
+  auto retry_executor_ = std::make_shared<folly::IOThreadPoolExecutor>(1);
   auto codec = std::make_shared<hbase::KeyValueCodec>();
   auto rpc_client = std::make_shared<RpcClient>(io_executor_, cpu_executor_, codec,
                                                 AsyncRpcRetryTest::test_util->conf());
