@@ -113,7 +113,7 @@ TEST(Result, FilledResult) {
   // Value will be nullptr as no such family and qualifier is present
   ASSERT_FALSE(result.Value("family-4", "qualifier"));
   // Value will be present as family and qualifier is present
-  ASSERT_TRUE(result.Value("family-4", "column-4") != none);
+  ASSERT_TRUE(result.Value("family-4", "column-4"));
   // Value should be present and match.
   EXPECT_EQ(latest_cell->Value(), (*result.ColumnLatestCell("family-4", "column-4")).Value());
   EXPECT_EQ("value-5", (*result.ColumnLatestCell("family-5", "column-5")).Value());
