@@ -20,7 +20,8 @@
 #include "hbase/client/time-range.h"
 #include "hbase/test-util/test-util.h"
 
-#include <glog/logging.h>
+#include <folly/GLog.h>
+#include <gtest/gtest.h>
 
 using namespace hbase;
 
